Update bucmaps
bucmaps:from_list/[1, 2] and bucmaps:to_list/[1, 2]

The transformation from a map to a list (or vice versa) is done by traversing all the elements in depth and width. Thus in the case of the transformation of a list to a map, all the lists, whatever their position in a list, a tuple, ... is transformed into a map.